renamed Section to TilesSection
