--- src/common/Section/index.js
@@ -1,11 +1,0 @@
-import { SectionTitle, StyledSection } from "./styled";
-
-const Section = ({ title, children }) => {
-  return (
-    <StyledSection>
-      <SectionTitle>{title}</SectionTitle>
-      {children}
-    </StyledSection>
-  );
-};
-export default Section;
