--- src/common/Section/styled.js
@@ -1,28 +1,0 @@
-import styled from "styled-components";
-
-const bpMobile = ({ theme }) => theme.breakpoint.mobile;
-
-export const StyledSection = styled.section`
-    margin: 56px auto 0;
-    max-width: 1368px;
-   
-    @media (max-width: ${bpMobile}px) {
-        padding: 24px 16px 0 16px;
-        max-width: 640px;
-  }
-`;
-
-export const SectionTitle = styled.h2`
-    margin: 0 0 24px 0;
-    color: ${({ theme }) => theme.color.woodsmoke};
-    font-weight: 600;
-    font-size: 36px;
-    line-height: 120%;
-
-    @media (max-width: ${bpMobile}px) {
-        margin: 0 0 12px 0;
-        max-width: 1368px;
-        font-size: 18px;
-
-  }
-`;
