+++ src/common/TilesSection/index.js
@@ -1,0 +1,11 @@
+import { SectionTitle, StyledTilesSection, SectionList } from "./styled";
+
+const TilesSection = ({ title, children }) => {
+  return (
+    <StyledTilesSection>
+      <SectionTitle>{title}</SectionTitle>
+      <SectionList>{children}</SectionList>
+    </StyledTilesSection>
+  );
+};
+export default TilesSection;
